Refactor theme structure
