+++ frontend/src/theme/lightTheme.ts
@@ -1,0 +1,9 @@
+import { ThemeOptions } from '@mui/material/styles';
+
+const lightThemeOptions: ThemeOptions = {
+  palette: {
+    mode: 'light',
+  },
+};
+
+export default lightThemeOptions;
